refactor: null 예외 처리 변경
