+++ src/main/resources/static/docs/coupon.html
@@ -627,7 +627,7 @@ Content-Length: 3
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2023-11-15 17:55:56 +0900
+Last updated 2023-11-15 18:24:57 +0900
 </div>
 </div>
 </body>
