+++ src/test/java/com/moabam/api/domain/repository/TokenRepostiroyTest.java
@@ -8,7 +8,6 @@ import java.time.Duration;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -40,40 +39,21 @@ class TokenRepostiroyTest {
 			.isThrownBy(() -> tokenRepository.saveToken(1L, TokenSaveValueFixture.tokenSaveValue()));
 	}
 
-	@Nested
-	@DisplayName("토큰 조회")
-	class Search {
+	@DisplayName("토큰 조회 성공")
+	@Test
+	void token_get_success() {
+		// given
+		willReturn(TokenSaveValueFixture.tokenSaveValue("token"))
+			.given(hashTemplateRepository).get(anyString());
 
-		@DisplayName("성공")
-		@Test
-		void token_get_success() {
-			// given
-			willReturn(TokenSaveValueFixture.tokenSaveValue("token"))
-				.given(hashTemplateRepository).get(anyString());
+		// when
+		TokenSaveValue tokenSaveValue = tokenRepository.getTokenSaveValue(123L);
 
-			// when
-			TokenSaveValue tokenSaveValue = tokenRepository.getTokenSaveValue(123L);
-
-			// then
-			assertAll(
-				() -> assertThat(tokenSaveValue).isNotNull(),
-				() -> assertThat(tokenSaveValue.refreshToken()).isEqualTo("token")
-			);
-		}
-
-		@DisplayName("실패")
-		@Test
-		void token_get_fail() {
-			// given
-			willReturn(null)
-				.given(hashTemplateRepository).get(anyString());
-
-			// when
-			TokenSaveValue tokenSaveValue = tokenRepository.getTokenSaveValue(123L);
-
-			// then
-			assertThat(tokenSaveValue).isNull();
-		}
+		// then
+		assertAll(
+			() -> assertThat(tokenSaveValue).isNotNull(),
+			() -> assertThat(tokenSaveValue.refreshToken()).isEqualTo("token")
+		);
 	}
 
 	@DisplayName("토큰 저장 삭제")
